Move fixed-point datatypes to their own package.
This adds a _little_ more friction in order to encourage the use of type
class constraints rather than hard-coding a particular type.
