+++ CHANGELOG/breaking.md
@@ -1,0 +1,1 @@
+- Move fixed-point datatypes to their own package.
