+++ core/jvm/src/test/scala/matryoshka/exp/package.scala
@@ -16,6 +16,8 @@
 
 package matryoshka
 
+import matryoshka.data.Fix
+
 import scala.{Int, Symbol}
 
 package object exp {
