+++ core/jvm/src/test/scala/matryoshka/exp2/package.scala
@@ -16,6 +16,8 @@
 
 package matryoshka
 
+import matryoshka.data.Fix
+
 import scala.Int
 
 package object exp2 {
