+++ core/jvm/src/test/scala/matryoshka/patterns/diff.scala
@@ -17,6 +17,7 @@
 package matryoshka.patterns
 
 import matryoshka._, Recursive.ops._
+import matryoshka.data.Mu
 import matryoshka.example._
 import matryoshka.exp._
 import matryoshka.helpers._
