+++ core/shared/src/main/scala/matryoshka/data/Nu.scala
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 
-package matryoshka
+package matryoshka.data
+
+import matryoshka._
 
 import scalaz._, Scalaz._
